[REFACTOR] : useLocalStorage hook 삭제, 안 쓰는 파일 삭제
