+++ src/pages/SignUp/index.tsx
@@ -1,9 +1,9 @@
 import { Navigate, useOutletContext } from 'react-router-dom';
 
 export default function SignUpPage() {
-  const authState = useOutletContext();
+  const isLoggedIn = useOutletContext();
 
-  if (authState) {
+  if (isLoggedIn) {
     return <Navigate to="/todo" />;
   }
   return <div>회원가입페이지</div>;
