+++ src/pages/Todo/index.tsx
@@ -1,9 +1,9 @@
 import { Navigate, useOutletContext } from 'react-router-dom';
 
 export default function TodoPage() {
-  const authState = useOutletContext();
+  const isLoggedIn = useOutletContext();
 
-  if (!authState) {
+  if (!isLoggedIn) {
     return <Navigate to="/signin" />;
   }
   return <div>투두페이지</div>;
