+++ src/routes/_globalLayout.tsx
@@ -3,10 +3,10 @@ import { Outlet } from 'react-router-dom';
 import useCheckAuth from '@/hooks/useCheckAuth';
 
 export default function Layout() {
-  const authState = useCheckAuth();
+  const isLoggedIn = useCheckAuth();
   return (
     <Suspense fallback="loading...">
-      <Outlet context={authState} />
+      <Outlet context={isLoggedIn} />
     </Suspense>
   );
 }
