+++ src/services/apiClient.tsx
@@ -1,5 +1,5 @@
 import axios, { AxiosInstance } from 'axios';
-import { BASE_URL } from '@/utils/config';
+import { BASE_URL } from '@/constants/config';
 
 const apiClient: AxiosInstance = axios.create({
   baseURL: BASE_URL,
