--- src/utils/config.ts
@@ -1,13 +1,0 @@
-export const BASE_URL = 'https://www.pre-onboarding-selection-task.shop';
-
-export const API_URL = {
-  signIn: '/auth/signin',
-  signUp: '/auth/signup',
-  todos: '/todos',
-};
-
-export const ROUTE_PATH = {
-  signIn: '/signin',
-  signUp: '/signup',
-  todo: '/todo',
-};
